api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pandas>=1.5.3
numpy>=1.23.5
scikit-learn>=1.2.2
joblib>=1.2.0
python-dotenv>=0.21.1
faiss-cpu==1.7.4
torch>=2.2.0
transformers==4.30.2
sentence-transformers==2.2.2
huggingface_hub==0.15.1